Rename Java package for Gradle subproject
