+++ gradle-with-plugin/src/main/amp/config/example-dynamic-extensions-module/log4j.properties
@@ -1,0 +1,1 @@
+log4j.logger.eu.xenit.de.example=INFO
