+++ gradle-with-plugin/src/main/java/example/Constants.java
@@ -1,0 +1,9 @@
+package example;
+
+class Constants {
+
+    final static String EXAMPLE_WEBSCRIPTS_FAMILY = "Example WebScripts";
+    final static String EXAMPLE_WEBSCRIPTS_BASE_URI = "/dynamic-extensions/examples";
+
+
+}
