+++ gradle-with-plugin/src/main/java/example/ExampleWorkflowTask.java
@@ -1,4 +1,4 @@
-package com.github.dynamicextensionsalfresco.examples;
+package example;
 
 import org.activiti.engine.delegate.DelegateExecution;
 import org.activiti.engine.delegate.JavaDelegate;
@@ -6,11 +6,13 @@ import org.springframework.stereotype.Component;
 
 /**
  * The component name must match the delegate expression in the workflow definition.
+ *
  * @author Laurent Van der Linden
  */
 @Component("example.dynamicextensions.testTask")
 public class ExampleWorkflowTask implements JavaDelegate {
-    public void execute(DelegateExecution execution) throws Exception {
+
+    public void execute(DelegateExecution execution) {
         execution.setVariable("wf_review_comment", "Processed by ServiceTask");
     }
 }
