+++ gradle-with-plugin/src/main/resources/example/TemplateWebScript.handleHelloTemplate.get.html.ftl
