+++ gradle-with-plugin/src/main/resources/example/templates/html.inc.ftl
